Add files via upload
E-DOC MANAGEMENT is a system that more similar with student University portal system. we have seen a lot of online system that make our student and lecturer life easier. But E-DOC MANAGEMENT is the first time to be created for the students to upload their medical certificate and other attachments for their respective subjects. On the other hand, the lecturers can inspect the student’s documents in their page too. A student who first time using this system must register their details first. And then they have to register for subjects listed in the system first. After they register for the subjects, the system will direct them to list of subjects they have registered in. so this can help the students to check their subjects’ details. And then they can proceed with uploading any documents they wanted to attach with. The student only can attach one document for each subject. If the student has more than one document the student has to register again for the same subject to add one more subject. This flow helps the lecturer to not get confuse with the documents. The students have to give their full details in the registration form. They can’t register if they leave any of the space with blank.